+++ app/Livewire/Admin/Modifiers.php
@@ -87,6 +87,8 @@ class Modifiers extends Component
 
     public function saveGroup(): void
     {
+        $this->authorize($this->editingGroupId ? 'edit_modifiers' : 'create_modifiers');
+
         $this->validate(['groupName' => 'required|min:2|max:100']);
 
         $data = [
@@ -109,6 +111,8 @@ class Modifiers extends Component
 
     public function deleteGroup(int $id): void
     {
+        $this->authorize('delete_modifiers');
+
         ModifierGroup::findOrFail($id)->delete();
         $this->dispatch('notify', type: 'success', message: 'Grup modifier berhasil dihapus');
     }
@@ -142,6 +146,8 @@ class Modifiers extends Component
 
     public function saveModifier(): void
     {
+        $this->authorize($this->editingModifierId ? 'edit_modifiers' : 'create_modifiers');
+
         $this->validate(['modifierName' => 'required|min:2|max:100']);
 
         $data = [
@@ -165,6 +171,8 @@ class Modifiers extends Component
 
     public function deleteModifier(int $id): void
     {
+        $this->authorize('delete_modifiers');
+
         Modifier::findOrFail($id)->delete();
         $this->dispatch('notify', type: 'success', message: 'Modifier berhasil dihapus');
     }
